+++ lib/Postfilter/ReleaseManager.pm
@@ -21,7 +21,7 @@ use File::Spec;
 use JSON::PP;
 use POSIX qw(strftime);
 
-our $VERSION = '2026.09.1-rc2';
+our $VERSION = '2026.09.1-rc3';
 
 use Exporter qw(import);
 our @EXPORT_OK = qw(